requirements.txt
open3d==0.8.0.0
PyYAML==5.3.1
tqdm==4.28.1
tensorboardX==1.2
termcolor==1.1.0
scikit-image==0.14.2
opencv-contrib-python>=3.4.2.17
opencv-python>=3.4.2.17,<4
imageio==2.3.0
trimesh==3.8.15
plyfile==0.6
PyMCubes==0.1.0
pyglet==1.4.0b1
chumpy
pytorch-ignite==0.2.0
yacs
apex
cmake==3.13.0
imgaug
lpips
tensorboard